#38
- src убран из файлов пакета для облегчения (patch)
- preset/bun: healthCheck перенесен в preset/isomorphic (minor)
