+++ package.json
@@ -7,9 +7,77 @@
   "engines": {
     "node": ">=16.15.1"
   },
+  "scripts": {
+    "prepare": "husky install",
+    "type-check": "npx tsc -p . --noEmit && npx tsc --project tsconfig.bun.json --noEmit",
+    "build:before": "rm -rf dist cache config di http log preset utils",
+    "build:esm": "babel src --config-file ./babel.config.esm.js -d dist/esm --source-maps --extensions '.ts,.tsx' --ignore='**/__test__/**/*' --ignore='**/*.d.ts' && node scripts/make-pkg-esm.mjs",
+    "build:cjs": "babel src --config-file ./babel.config.cjs.js -d dist/cjs --source-maps --extensions '.ts,.tsx' --ignore='**/__test__/**/*' --ignore='**/*.d.ts' && node scripts/make-pkg-cjs.mjs",
+    "build:types": "tsc --project tsconfig.types.json",
+    "post-build-esm": "node scripts/post-build-esm.mjs",
+    "build": "npm run build:before && npm run build:esm && npm run build:cjs && npm run build:types && npm run post-build-esm",
+    "lint": "eslint --cache ./src --ext .js,.jsx,.ts,.tsx",
+    "test": "jest"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.23.0",
+    "@babel/core": "^7.23.0",
+    "@babel/preset-env": "^7.22.5",
+    "@babel/preset-react": "^7.22.5",
+    "@babel/preset-typescript": "^7.22.5",
+    "@sima-land/linters": "^4.0.0",
+    "@testing-library/react": "^14.0.0",
+    "@types/express": "^4.17.17",
+    "@types/jest": "^28.1.7",
+    "@types/jsesc": "^3.0.1",
+    "@types/react": "^18.2.14",
+    "@types/react-dom": "^18.2.6",
+    "@types/webpack": "^5.28.1",
+    "babel-jest": "^29.5.0",
+    "bun-types": "^1.0.12",
+    "eslint": "^8.53.0",
+    "fs-extra": "^11.1.1",
+    "husky": "^8.0.3",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "prettier": "^3.0.3",
+    "typescript": "^5.3.2",
+    "whatwg-fetch": "^3.6.17"
+  },
+  "dependencies": {
+    "@humanwhocodes/env": "^2.2.2",
+    "@krutoo/fetch-tools": "^0.0.11",
+    "@opentelemetry/api": "^1.4.1",
+    "@opentelemetry/exporter-prometheus": "^0.38.0",
+    "@opentelemetry/exporter-trace-otlp-http": "^0.39.1",
+    "@opentelemetry/propagator-jaeger": "^1.12.0",
+    "@opentelemetry/resources": "^1.12.0",
+    "@opentelemetry/sdk-metrics": "^1.12.0",
+    "@opentelemetry/sdk-trace-base": "^1.12.0",
+    "@opentelemetry/sdk-trace-node": "^1.12.0",
+    "@opentelemetry/semantic-conventions": "^1.12.0",
+    "@sentry/browser": "^7.81.0",
+    "@sentry/bun": "^7.81.0",
+    "@sentry/node": "^7.81.0",
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "jsesc": "^3.0.2",
+    "middleware-axios": "^2.1.5",
+    "pino": "^8.14.1",
+    "pino-pretty": "^9.1.1",
+    "prom-client": "^14.2.0",
+    "webpack": "^5.82.0"
+  },
+  "peerDependencies": {
+    "@reduxjs/toolkit": "^1.9.5",
+    "axios": "^1.4.0",
+    "react": "^17.0 || ^18.0",
+    "react-dom": "^17.0 || ^18.0",
+    "react-redux": "^8.0.5",
+    "redux-saga": "^1.2.3"
+  },
   "files": [
     "dist",
-    "src",
     "README.md"
   ],
   "exports": {
@@ -260,74 +328,5 @@
         "./dist/types/utils/webpack/index.d.ts"
       ]
     }
-  },
-  "scripts": {
-    "prepare": "husky install",
-    "type-check": "npx tsc -p . --noEmit && npx tsc --project tsconfig.bun.json --noEmit",
-    "build:before": "rm -rf dist cache config di http log preset utils",
-    "build:esm": "babel src --config-file ./babel.config.esm.js -d dist/esm --source-maps --extensions '.ts,.tsx' --ignore='**/__test__/**/*' --ignore='**/*.d.ts' && node scripts/make-pkg-esm.mjs",
-    "build:cjs": "babel src --config-file ./babel.config.cjs.js -d dist/cjs --source-maps --extensions '.ts,.tsx' --ignore='**/__test__/**/*' --ignore='**/*.d.ts' && node scripts/make-pkg-cjs.mjs",
-    "build:types": "tsc --project tsconfig.types.json",
-    "post-build-esm": "node scripts/post-build-esm.mjs",
-    "build": "npm run build:before && npm run build:esm && npm run build:cjs && npm run build:types && npm run post-build-esm",
-    "lint": "eslint --cache ./src --ext .js,.jsx,.ts,.tsx",
-    "test": "jest"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.23.0",
-    "@babel/core": "^7.23.0",
-    "@babel/preset-env": "^7.22.5",
-    "@babel/preset-react": "^7.22.5",
-    "@babel/preset-typescript": "^7.22.5",
-    "@sima-land/linters": "^4.0.0",
-    "@testing-library/react": "^14.0.0",
-    "@types/express": "^4.17.17",
-    "@types/jest": "^28.1.7",
-    "@types/jsesc": "^3.0.1",
-    "@types/react": "^18.2.14",
-    "@types/react-dom": "^18.2.6",
-    "@types/webpack": "^5.28.1",
-    "babel-jest": "^29.5.0",
-    "bun-types": "^1.0.12",
-    "eslint": "^8.53.0",
-    "fs-extra": "^11.1.1",
-    "husky": "^8.0.3",
-    "jest": "^28.1.3",
-    "jest-environment-jsdom": "^28.1.3",
-    "prettier": "^3.0.3",
-    "typescript": "^5.3.2",
-    "whatwg-fetch": "^3.6.17"
-  },
-  "dependencies": {
-    "@humanwhocodes/env": "^2.2.2",
-    "@krutoo/fetch-tools": "^0.0.11",
-    "@opentelemetry/api": "^1.4.1",
-    "@opentelemetry/exporter-prometheus": "^0.38.0",
-    "@opentelemetry/exporter-trace-otlp-http": "^0.39.1",
-    "@opentelemetry/propagator-jaeger": "^1.12.0",
-    "@opentelemetry/resources": "^1.12.0",
-    "@opentelemetry/sdk-metrics": "^1.12.0",
-    "@opentelemetry/sdk-trace-base": "^1.12.0",
-    "@opentelemetry/sdk-trace-node": "^1.12.0",
-    "@opentelemetry/semantic-conventions": "^1.12.0",
-    "@sentry/browser": "^7.81.0",
-    "@sentry/bun": "^7.81.0",
-    "@sentry/node": "^7.81.0",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "jsesc": "^3.0.2",
-    "middleware-axios": "^2.1.5",
-    "pino": "^8.14.1",
-    "pino-pretty": "^9.1.1",
-    "prom-client": "^14.2.0",
-    "webpack": "^5.82.0"
-  },
-  "peerDependencies": {
-    "@reduxjs/toolkit": "^1.9.5",
-    "axios": "^1.4.0",
-    "react": "^17.0 || ^18.0",
-    "react-dom": "^17.0 || ^18.0",
-    "react-redux": "^8.0.5",
-    "redux-saga": "^1.2.3"
   }
 }
